Move plot scripts inside package
