+++ requirements.txt
@@ -7,7 +7,6 @@ gym-minigrid
 scikit-optimize
 optuna
 pytablewriter~=0.64
-seaborn
 pyyaml>=5.1
 cloudpickle>=1.5.0
 plotly
