+++ rl_zoo3/plots/__init__.py
@@ -1,0 +1,3 @@
+from rl_zoo3.plots.all_plots import all_plots
+from rl_zoo3.plots.plot_from_file import plot_from_file
+from rl_zoo3.plots.plot_train import plot_train
