+++ setup.cfg
@@ -15,11 +15,11 @@ inputs = .
 ignore = W503,W504,E203,E231  # line breaks before and after binary operators
 # Ignore import not used when aliases are defined
 per-file-ignores =
-	./rl_zoo3/__init__.py:F401
+  ./rl_zoo3/__init__.py:F401
+  ./rl_zoo3/plots/__init__.py:F401
   ./rl_zoo3/import_envs.py:F401
-  ./scripts/all_plots.py:E501
-  ./scripts/plot_train.py:E501
-  ./scripts/plot_training_success.py:E501
+  ./rl_zoo3/plots/all_plots.py:E501
+  ./rl_zoo3/plots/plot_train.py:E501
 
 exclude =
     # No need to traverse our git directory
